requirements.txt
streamlit
PyQt6
pyqtgraph
websocket-client
bleak
pandas
numpys
plotly
pyserial 